+++ src/main/java/net/neganote/gtutilities/config/UtilConfig.java
@@ -29,7 +29,7 @@ public class UtilConfig {
 
         @Configurable
         @Configurable.Comment({ "Whether the Sterile Cleaning Maintenance Hatch is enabled." })
-        public boolean sterileHatchEnabled = true;
+        public boolean sterileHatchEnabled = false;
         @Configurable
         @Configurable.Comment({ "Whether the 64A energy converters are enabled." })
         public boolean converters64aEnabled = true;
@@ -45,11 +45,11 @@ public class UtilConfig {
         public long omnibreakerEnergyCapacity = 40_960_000L;
 
         @Configurable
-        @Configurable.Comment({ "Whether the Power-Transfer Einstein-Rosen Bridge is enabled." })
+        @Configurable.Comment({ "Whether the Wireless Active Transformer is enabled." })
         public boolean pterbEnabled = true;
 
         @Configurable
-        @Configurable.Comment({ "Base amount of PTERB coolant to drain every second.",
+        @Configurable.Comment({ "Base amount of WAT coolant to drain every second.",
                 "(Setting both this amount and the IO multiplier to 0 disables the coolant mechanic.)" })
         public int pterbCoolantBaseDrain = 0;